ReportDetailDTO 추가, 예외처리, 검증 추가(책임자인지, 보고 관련자인지 등등), 보고 완료상태 수정, 상세보고 작성 기능 추가, 맵퍼 테스트 완료, 포스트맨 전부 정상 작동 확인,

+++ src/main/java/com/root34/aurora/report/dto/ReportDTO.java
@@ -18,7 +18,7 @@ public class ReportDTO {
     /**
      * @variable memberCode 멤버 코드
      */
-    private long memberCode;
+    private int memberCode;
     /**
      * @variable reportType 보고 유형
      */
